fix: move pancakeswap2 to pancakeswap.info api
For Pancakeswap 1 we used pancakeswap.info, but they no longer support
it. They moved to the v2 exchange, and their rest api is much more
dev-friendly then the combined apis we currentyly use for Pancakeswap 2,
of which one (streamingfast) currently does not work anymore
