+++ drivers/index.js
@@ -254,7 +254,6 @@ exports.Orebz = require('./orebz');
 exports.Osmosis = require('./osmosis');
 exports.Ovex = require('./ovex');
 exports.P2pb2b = require('./p2pb2b');
-exports.Pancakeswap = require('./pancakeswap');
 exports.Pancakeswap2 = require('./pancakeswap2');
 exports.Paribu = require('./paribu');
 exports.Parsxc = require('./parsxc');
